safari-600.7-branch/LayoutTests/js/dfg-int52-spill-expected.txt

Tests that spilling an int52 works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(2000000000, array) is 209000001908 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
